requirements.txt
build==1.2.2.post1
click==8.2.1
evdev==1.9.2
packaging==25.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
